Off message fix
When the BLE command tells the HDS to power off, it often just blink a little and no Off message is shown.
b_powerOff was added, and when it was set to true, the loop() will check it and goes off.

This bug is because when inside a u8g2 cycle, another u8g2 cycle was called by BLE command. It should wait for the current cycle to complete then show the Off message.

Also, the shut_down_now_nobeep called in setup() was also changed to esp32_sleep(), because the nobeep function contains u8g2 code, which was never inited.